Extract babel-plugin-transform-react into a separate file
