+++ packages/babel-plugin-extract-messages/test/fixtures/jsx/integration-with-macro.js
@@ -1,0 +1,25 @@
+import { Trans } from "@lingui/react-macro"
+
+;<Trans>Hi, my name is {name}</Trans>
+;<span title={i18n.t`Title`} />
+;<span
+  title={i18n.plural({
+    value: count,
+    one: "# book",
+    other: "# books"
+  })}
+/>
+
+const a = i18n.t`Title`
+i18n.t`Title`
+
+const p = i18n.plural({
+  value: count,
+  one: "# book",
+  other: "# books"
+})
+i18n.plural({
+  value: count,
+  one: "# book",
+  other: "# books"
+})
