+++ packages/babel-plugin-transform-react/index.js
@@ -1,2 +1,2 @@
-// Can't use export here as the babel macro may not have this enabled
-module.exports = require("./src")
+export { default } from "./src"
+export { default as implementationPlugin } from "./src/implementationPlugin"
